Reduce test output verbosity
To better match the expectations for NUnit test output, eliminate trace
describing passed tests when running from the NUnit console. Failed test
info will still appear in the console. Note that ReSharper will still
display the full trace for individual specs in the test results window,
which is reasonable because it will only show the output for the single
selected spec.