GenericSpec: Ensure the MockingContainer is always properly initialized
It is somewhat common for spec writers to override BeforeEachExample, but forget
to call base.BeforeEachExample. If the spec being written inherits from
GenericSpec, this would result in a somewhat obscure exception about the mocking
container not being initialized. This is because the mocking container was
previously initialized via GenericSpec's BeforeEachExample.

This commit refactors how GenericSpec plugs into Spec's Before pipeline such
that the initialization of the mocking container always happens, regardless of
what happens in BeforeEachExample.

To (hopefully) make this clearer the specs in SpecEasy.Specs have been updated
to remove their now-unnecessary calls to base.(Before|After)EachExample.